// src/lib/klayr/types/klayr.ts
export interface Transaction {
  readonly module: string
  readonly command: string
  readonly senderPublicKey: Buffer
  readonly nonce: bigint
  readonly fee: bigint
  readonly params: Buffer
  readonly signatures: ReadonlyArray<Buffer>
  readonly id: Buffer
}

export type TransactionJSON = JSONObject<Transaction>
export type DecodedTransaction<T = Record<string, unknown>> = Omit<Transaction, 'params'> & {
  params: T
}